requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
filters==1.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
